authentication with user id

+++ controllers/contactController.js
@@ -48,6 +48,11 @@ const updateContact = asyncHandler(async (req,res)=>{
 		res.status(404);
 		throw new Error("Contact not found");
 	}
+
+	if(contact.user_id.toString() !== req.user.id){
+		res.status(403);
+		throw new Error("User don't have permission to update other user contacts")
+	}
 	const update = await Contact.findByIdAndUpdate(
 		req.params.id,
 		req.body,
@@ -65,6 +70,12 @@ const deleteContact = asyncHandler(async (req,res)=>{
 		res.status(404);
 		throw new Error("Contact not found");
 	}
+
+	if(contact.user_id.toString() !== req.user.id){
+		res.status(403);
+		throw new Error("User don't have permission to delete other user contacts")
+	}
+
 	await Contact.findByIdAndDelete(req.params.id);
 	res.status(200).json(contact);
 });
